Radius can grow

+++ app/src/main/java/model/coordinate/CoordinateInteger.java
@@ -14,12 +14,12 @@ public class CoordinateInteger extends Coordinate<Integer,Direction> {
     }
 
     @Override
-    public CoordinateInteger placeCoordinateFrom(Direction direction, Integer distance) {
+    public CoordinateInteger placeCoordinateFrom(Direction direction, double distance) {
        switch (direction) {
-        case UP:return new CoordinateInteger(this.x, this.y - distance);
-        case DOWN:return new CoordinateInteger(this.x, this.y + distance);
-        case LEFT:return new CoordinateInteger(this.x - distance, this.y);
-        case RIGHT:return new CoordinateInteger(this.x + distance, this.y);
+        case UP:return new CoordinateInteger(this.x, (int) (this.y - distance));
+        case DOWN:return new CoordinateInteger(this.x, (int) (this.y + distance));
+        case LEFT:return new CoordinateInteger((int) (this.x - distance), this.y);
+        case RIGHT:return new CoordinateInteger((int) (this.x + distance), this.y);
         default:
             throw new IllegalArgumentException("The direction " + direction + " is not a valid direction");
        }
